Skyline: Merge SpilledChromatograms for real this time.
Former-commit-id: 73415258a9f53416a07818ff6b73f01e7f5f1cca

+++ pwiz_tools/Skyline/Model/Results/RemoteApi/ChromatogramGeneratorTask.cs
@@ -135,7 +135,7 @@ namespace pwiz.Skyline.Model.Results.RemoteApi
             times = tranInfo.Times;
             if (null != tranInfo.ScanIds)
             {
-                scanIds = tranInfo.ScanIds[(int) chromKeyIndices.Key.Source];
+                scanIds = tranInfo.ScanIds[(short) chromKeyIndices.Key.Source];
             }
             else
             {
